requirements.txt
jinja2>=3.1.0
pyyaml>=6.0
typer>=0.9.0
pydantic>=2.0.0
